requirements.txt
python-telegram-bot==20.0a4
aioschedule==0.5.2
requests==2.22.0
beautifulsoup==4.11.1
python-dotenv==0.21.0
